requirements.txt
Babel==2.3.4
colorlog==2.7.0
debtcollector==1.5.0
decorator==4.0.10
eventlet==0.19.0
flake8==2.5.4
flake8-double-quotes==0.0.1
flake8-import-order==0.8
funcsigs==1.0.2
greenlet==0.4.9
mccabe==0.4.0
msgpack-python==0.4.7
netaddr==0.7.18
networkx==1.11
oslo.config==3.11.0
oslo.i18n==3.7.0
pbr==1.10.0
pep8==1.7.0
pep8-naming==0.3.3
pyflakes==1.0.0
pytz==2016.4
PyYAML==3.11
repoze.lru==0.6
Routes==2.3.1
ryu==4.3
six==1.10.0
stevedore==1.15.0
WebOb==1.6.1
wrapt==1.10.8
